(!see notes!) fix dot assign/null/calcClassStorage
This is a monster commit (sorry!). The first thing this commit does is
change the `default` behavior in calculateClassStorageNeeds to store a
pointer to the class instead of an i32 pointer; the i32 is a remnant
from before I started understanding how LLVM is dealing with objects.
This change necessitated a change in DJProgram::codeGen(): we now
populate allocatedClasses *before* determining the class storage sizes.
This is so that the pointers can be allocated appropriately for a class
which has, e.g., a pointer to another class or itself as a field.

This commit also adds the default `type` argument to every codeGen
method. This is only used in DJNull, to resolve the issue I've been
having with expressions with null children: LLVM requires null to be
well-typed when the IR is generated. The way I handled it is a little
hacky (more on that in the next commit,) but it does seem to work so
far.

I also changed the return value for DJDotAssign to be more in line with
the actual return semantics of DJ.