add null special variables to DJExpression
This commit adds some extra information to DJExpression nodes, which
will be used when one of these nodes has a NULL_EXPR as a child. (See
last commit for why this extra type information is needed.) The only one
that is tested to work right now is DOT_ASSIGN_EXPR; equality
expresisons are certainly broken.