Fix null for equality expr
Another big commit! This one fixes null expression code generation when on
either sides of an equality expression. We leverage the type-checker to
set staticClassNum when exactly one of an equality expression's children
is a NULL_EXPR. During code generation, we use this as the argument to
DJNull::codeGen to return a properly-typed null.